Fixed bugs in the script, and rewrote it to work with any mode number.
TODO: Try to find a fix so that it is not necessary to read the whole
file into memory at once. Iterating over it is not doable now as I rely
on splitting by "Mode Number". Maybe all that is needed is to find the
length of the first "Mode data" and then use that.